requirements.txt
selenium==4.15.2
webdriver-manager==4.0.1
textblob==0.17.1
pandas==2.1.0
matplotlib==3.7.1
pyyaml==6.0.1
fake-useragent==1.4.0
python-dotenv==1.0.0
requests==2.31.0 